+++ docs/changelog.rst
@@ -12,7 +12,7 @@ v0.15.0 (UNRELEASED)
 **Dependencies**
 
 - Mopidy no longer supports Python 2.6. Currently, the only Python version
-  supported by Mopidy is Python 2.7.
+  supported by Mopidy is Python 2.7. (Fixes: :issue:`344`)
 
 
 v0.14.1 (2013-04-28)